Fix recognition of the letter I
OCR cannot seem to find it, presumably because it looks too much like
a plain rectangle when on its own.

A separate model is trained to recognize just that letter (from the
cropped and cleaned up black and white image), and that object detection
result is used to augment the OCR results (all of this is in refactored
BoardLetterExtractor class).

Conceivably this approach could be extended to other letters that may
be troublesome, the object label is the letter and the model could be
trained on more of them.