Switch to doing own solution generation
https://github.com/imjasonh/spelling-bee only has solutions for the
print version, where words must be at least 5 letters long. We instead
embed a dictioanry into the app and traverse it to find candidates.

The dictionary was generated by taking the the 1/3 million most frequent
words from the Google Web Trillion Word Corpus (count_1w.txt from
https://norvig.com/ngrams/) and then filtering it to include only
words with at least 4 letters, with no more than 7 unique letters, and
to only consider words that are correctly spelled according to aspell.
This still ends up with a lot of proper nouns and other words that are
not allowed, but is still better than nothing.

This has the nice side effect that words are sorted by their commonality,
so solutions are presented in order of "obviousness"